+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v0.66.0
+## v0.65.2
 
 ### Fixed